chore: Update description and authors in pyproject.toml
This commit updates the description and authors fields in the pyproject.toml file for the "dhti-elixir-base" package. The description is changed to "Base classes for DHTI Elixir" and the authors field is updated to include "Bell Eapen <github.com/dermatologist>". This change provides more accurate and informative information about the package.